Move HandlerConfig to common/server/ for sharing across apps
