+++ source/lib/common/server/HandlerConfig.js
