+++ source/lib/server/Router.js
@@ -1,7 +1,7 @@
 import path from 'path';
 import { fileURLToPath } from 'url';
 import express from 'express';
-import { HandlerConfig } from './HandlerConfig.js';
+import { HandlerConfig } from '../common/server/HandlerConfig.js';
 import { AssetsHandlerExecutor } from './handlers/AssetsHandlerExecutor.js';
 import { EngineContinueHandlerExecutor } from './handlers/engine/EngineContinueHandlerExecutor.js';
 import { EnginePauseHandlerExecutor } from './handlers/engine/EnginePauseHandlerExecutor.js';
